% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niivuer.R
\name{niivue}
\alias{niivue}
\title{'NiiVue' for R}
\usage{
niivue(message, width = NULL, height = NULL, elementId = NULL)
}
\description{
Creates a widget that bridges 'NiiVue', a powerful 3D brain viewer and R-shiny
}
