% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niivuer.R
\name{niivuer}
\alias{niivuer}
\title{'NiiVue' for R}
\usage{
niivuer(
  volume_list = list(list(url = "https://niivue.github.io/niivue/images/mni152.nii.gz",
    colormap = "gray"), list(url = "https://niivue.github.io/niivue/images/hippo.nii.gz",
    colormap = "red")),
  mesh_list = list(list(url =
    "https://niivue.github.io/niivue/images/BrainMesh_ICBM152.lh.mz3")),
  viewer_settings = list(setMultiplanarLayout = 2, setSliceMM = TRUE, setSliceType = 4),
  controller_settings = list(),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{volume_list}{list of items, each item is a volume file}

\item{mesh_list}{list of items, each item is a mesh file}

\item{viewer_settings}{initial call to set viewer status}

\item{controller_settings}{initial controller settings}

\item{...}{application settings}

\item{width, height}{widget size}

\item{elementId}{passed to \code{\link[htmlwidgets]{createWidget}}.}
}
\value{
A \code{'niivuer'} widget
}
\description{
Creates a widget that bridges 'NiiVue', a powerful 3D brain viewer and R-shiny
}
\examples{

if( interactive() ) {

  niivuer()

}


}
